+++ lib/ruby_aem/resources/group.rb
@@ -48,12 +48,9 @@ module RubyAem
       # @return RubyAem::Result
       def delete()
         result = find_authorizable_id
-        if result.data
-          @call_params[:path] = RubyAem::Swagger.path(@call_params[:path])
-          @client.call(self.class, __callee__.to_s, @call_params)
-        else
-          result
-        end
+        @call_params[:authorizable_id] = result.data
+        @call_params[:path] = RubyAem::Swagger.path(@call_params[:path])
+        @client.call(self.class, __callee__.to_s, @call_params)
       end
 
       # Check whether the group exists or not.
@@ -62,6 +59,8 @@ module RubyAem
       #
       # @return RubyAem::Result
       def exists()
+        result = find_authorizable_id
+        @call_params[:authorizable_id] = result.data
         @call_params[:path] = RubyAem::Swagger.path(@call_params[:path])
         @client.call(self.class, __callee__.to_s, @call_params)
       end
@@ -83,12 +82,10 @@ module RubyAem
       # @return RubyAem::Result
       def add_member(member)
         result = find_authorizable_id
-        if result.data
-          @call_params[:member] = member
-          @client.call(self.class, __callee__.to_s, @call_params)
-        else
-          result
-        end
+        @call_params[:authorizable_id] = result.data
+        @call_params[:member] = member
+        @call_params[:path] = RubyAem::Swagger.path(@call_params[:path])
+        @client.call(self.class, __callee__.to_s, @call_params)
       end
 
       # Find the group's authorizable ID.